Finished combining the code
Used searchAllSongs JS in Dashboard UI
